Rename osd_config.h to options_osd.h for MPLAB-X projects.
