+++ RollPitchYaw/RollPitchYaw-udb5.X/nbproject/configurations.xml
@@ -103,9 +103,9 @@
         <itemPath>../../RollPitchYaw/Config/FSconfig.h</itemPath>
         <itemPath>../../RollPitchYaw/Config/nv_memory_options.h</itemPath>
         <itemPath>../../RollPitchYaw/Config/options.h</itemPath>
-        <itemPath>../../RollPitchYaw/Config/osd_config.h</itemPath>
         <itemPath>../../RollPitchYaw/Config/ports_config.h</itemPath>
         <itemPath>../../Config/options_magnetometer.h</itemPath>
+        <itemPath>../Config/options_osd.h</itemPath>
       </logicalFolder>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
@@ -244,6 +244,7 @@
     <Elem>../../MatrixPilot</Elem>
     <Elem>../../MAVLink</Elem>
     <Elem>../../Config</Elem>
+    <Elem>../Config</Elem>
   </sourceRootList>
   <projectmakefile>Makefile</projectmakefile>
   <confs>
