+++ Tools/LedTest/LedTest-udb4.X/nbproject/configurations.xml
@@ -87,9 +87,9 @@
         <itemPath>../../../Tools/LedTest/Config/FSconfig.h</itemPath>
         <itemPath>../../../Tools/LedTest/Config/nv_memory_options.h</itemPath>
         <itemPath>../../../Tools/LedTest/Config/options.h</itemPath>
-        <itemPath>../../../Tools/LedTest/Config/osd_config.h</itemPath>
         <itemPath>../../../Tools/LedTest/Config/ports_config.h</itemPath>
         <itemPath>../../../Config/options_magnetometer.h</itemPath>
+        <itemPath>../Config/options_osd.h</itemPath>
       </logicalFolder>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
@@ -213,6 +213,8 @@
     <Elem>../../MatrixPilot</Elem>
     <Elem>../../MAVLink</Elem>
     <Elem>../../../Config</Elem>
+    <Elem>../../../RollPitchYaw/Config</Elem>
+    <Elem>../Config</Elem>
   </sourceRootList>
   <projectmakefile>Makefile</projectmakefile>
   <confs>
